Fixes a bug in one of the tests for opposite rule break detection.
In the test run on test_op_break5_break_pos, the first point above the baseline mean is the first point after the end of the calculation period, i.e. the one at row 22. If we count the first opposite rule break occurring in the candidate period, comprising the first 10 points, then it isn't until the 4th point (row 25) that we get a candidate period with no opposite rule breaks.

However, on looking at this case, I believe that we should ignore the first run of the candiate period in checking for opposite rule breaks, since if it qualifies, it can only represent a transition on the way to the new process; in this case I think the recalculation should go ahead. So the new correct recalculation position is 22.